Add files via upload
The goal of this project is to create an embedded device using a Raspberry Pi that utilizes PWM (Pulse Width Modulation) to control the behavior of an LED. This device is designed to detect the presence of approaching objects using ultrasonic sensors hc- sr04 and provide user alerts. As an object comes closer to the device, the alert intensity increases through increased brightness of an LED. This GitHub repository contains the code for building and using this innovative object detection and alert system.